requirements.txt
# Core
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
streamlit>=1.25.0
requests>=2.31.0

# Configuration & validation
pydantic-settings>=2.1.1
python-dotenv>=1.1.0

# LLM SDKs
openai>=1.8.0
anthropic>=2024.4.0

# Testing (optional)
pytest>=7.4.0
